requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.3
asttokens==2.0.8
attrs==24.3.0
backcall==0.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
debugpy==1.6.3
decorator==5.1.1
dnspython==2.7.0
entrypoints==0.4
exceptiongroup==1.2.2
executing==1.0.0
fastapi==0.115.6
filelock==3.16.1
fsspec==2024.12.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
ipykernel==6.15.3
ipython==8.5.0
jedi==0.18.1
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-core==4.11.1
jupyter_client==7.3.5
llama2==0.0.1.dev0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.22.0
nest-asyncio==1.5.5
networkx==3.4.2
numpy==2.2.1
packaging==21.3
pandas==2.2.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
prompt-toolkit==3.0.31
proto-plus==1.25.0
protobuf==5.29.3
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.13.0
pymongo==4.10.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.2
pyzmq==24.0.1
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.5.0
starlette==0.41.3
streamlit==1.41.1
sympy==1.13.3
tenacity==9.0.0
toml==0.10.2
torch==2.2.2
tornado==6.2
tqdm==4.67.1
traitlets==5.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
utils==1.0.2
uvicorn==0.34.0
wcwidth==0.2.5
